requirements.txt
django="3.1.1"
psycopg2-binary
djangorestframework = "3.11.1"
authy
phonenumbers
python-dotenv
twilio
pyyaml
django-rest-swagger